+++ src/icp.h
@@ -11,7 +11,7 @@ extern "C" {
 #define MAX_ITERATIONS_KNOWN    5
 #define NUM_STANDARD_DEVS		1.5
 
-float icp(point4D *scan, node *root, float T[4][4], uint32_t numPts, uint8_t iterations);
+float icp(point4D *scan, node *root, float T[4][4], unsigned int numPts, unsigned int iterations);
 
 #ifdef __cplusplus
 }
